refactor: serializable event types refactored
