+++ src/input_device_monitor/event_caster/clone_caster/window_caster.rs
@@ -1,4 +1,4 @@
-use crate::input_device_monitor::my_event::MyWindowEvent;
+use crate::input_device_monitor::my_event::serializable_clone::MyWindowEvent;
 use nannou::event::WindowEvent;
 
 impl From<WindowEvent> for MyWindowEvent {
