+++ src/input_device_monitor/event_caster/mod.rs
@@ -1,4 +1,1 @@
-pub mod key_caster;
-pub mod mouse_button_caster;
-pub mod vec2_caster;
-pub mod window_caster;
+pub mod clone_caster;
