+++ src/input_device_monitor/my_event/flatbuffer/mod.rs
@@ -1,0 +1,5 @@
+pub mod flatbuffers_events;
+
+pub use flatbuffers_events::coffee_time::input_events::{
+    Event, EventArgs, Keyboard, KeyboardArgs, Mouse, MouseArgs, Position, PositionArgs,
+};
